Redesign fmVTScanResult
