+++ Widget/Forms/ThemeContainerControl.cs
@@ -8,7 +8,7 @@ namespace Widget.Forms
     abstract class ThemeContainerControl : ContainerControl
     {
 
-        #region " Initialization "
+        #region Initialization
 
         protected Graphics G;
         protected Bitmap B;
@@ -27,7 +27,7 @@ namespace Widget.Forms
 
         #endregion
 
-        #region " Convienence "
+        #region Convienence
 
         public abstract void PaintHook();
         protected override sealed void OnPaint(PaintEventArgs e)
@@ -88,57 +88,4 @@ namespace Widget.Forms
         #endregion
 
     }
-
-    class GroupBox : ThemeContainerControl
-    {
-
-        public GroupBox()
-        {
-            AllowTransparent();
-            _Border1 = new Pen(Color.FromArgb(90, Color.Black));
-            _Border2 = new Pen(Color.FromArgb(14, Color.White));
-        }
-
-        private Pen _Border1;
-        public Color Border1
-        {
-            get { return _Border1.Color; }
-            set
-            {
-                _Border1 = new Pen(value);
-                Invalidate();
-            }
-        }
-
-        private Pen _Border2;
-        public Color Border2
-        {
-            get { return _Border2.Color; }
-            set
-            {
-                _Border2 = new Pen(value);
-                Invalidate();
-            }
-        }
-
-        private Color _FillColor = Color.Transparent;
-        public Color FillColor
-        {
-            get { return _FillColor; }
-            set
-            {
-                _FillColor = value;
-                Invalidate();
-            }
-        }
-
-        public override void PaintHook()
-        {
-            G.Clear(_FillColor);
-            DrawBorders(_Border1, _Border2, ClientRectangle);
-            DrawCorners(BackColor, ClientRectangle);
-        }
-
-    }
-
 }
