+++ Widget/Forms/fmSettings.cs
@@ -95,6 +95,6 @@ namespace Widget
             return newKey;
         }
 
-  
+
     }
 }
